requirements.txt
openai
python-dotenv
langchain
fastapi
uvicorn
gunicorn
langchain-community